+++ flask_express/response.py
@@ -102,7 +102,7 @@ class Response(ResponseBase):
         if isinstance(content, (dict, list)):
             return self.json(content)
 
-        if isinstance(content, Munch) or issubclass(content, Munch): # for munch object.
+        if isinstance(content, Munch) or issubclass(content.__class__, Munch): # for munch object.
             return self.json(content.toDict())
 
         try:
@@ -564,6 +564,30 @@ class Response(ResponseBase):
         """
         return self.clear_cookie(*wargs, **kwargs)
 
+    def make_response_from_obj(self, rv:"Response", headers=None, status=None) -> "Response":
+        """
+        take a response object as the parameter and 
+        return the new response object from the old instance.
+
+        :param rv: the response object
+        """
+        self.status = rv.status
+        self.status_code = rv.status_code
+        self.headers = rv.headers
+        self.content_type = rv.content_type
+        self.direct_passthrough = rv.direct_passthrough
+
+        if headers is not None:
+            self.headers.update(headers)
+        
+        if status is not None:
+            if isinstance(status, (str, bytes, bytearray)):
+                self.status = status  # type: ignore
+            else:
+                self.status_code = status
+        
+        return self.make_response(rv.response)
+
 
     def make_response(self,
                 response: t.Optional[
@@ -589,7 +613,8 @@ class Response(ResponseBase):
         self.mimetype = mimetype or self.mimetype
         self.content_type = content_type or self.content_type
         self.direct_passthrough = direct_passthrough or self.direct_passthrough
-
+        # print ("data")
+        # print (self.data)
         return self.__class__(response=response, 
                 status=self.status, 
                 mimetype=self.mimetype, 
@@ -597,30 +622,6 @@ class Response(ResponseBase):
                 direct_passthrough=self.direct_passthrough, 
                 headers=self.headers
                 )
-
-    def make_response_from_obj(self, rv:"Response", headers=None, status=None) -> "Response":
-        """
-        take a response object as the parameter and 
-        return the new response object from the old instance.
-
-        :param rv: the response object
-        """
-        self.status = rv.status
-        self.status_code = rv.status_code
-        self.headers = rv.headers
-        self.content_type = rv.content_type
-        self.direct_passthrough = rv.direct_passthrough
-
-        if headers is not None:
-            self.headers.update(headers)
-        
-        if status is not None:
-            if isinstance(status, (str, bytes, bytearray)):
-                self.status = status  # type: ignore
-            else:
-                self.status_code = status
-        
-        return self.make_response(rv.response)
 
 
 from ._helper import Utils # added at bottom to solve the circular import issue.