+++ tests/conftest.py
@@ -7,13 +7,15 @@ import typing as t
 if t.TYPE_CHECKING:
     from flask.testing import FlaskClient
 
+BASE_DIR:str = os.path.abspath(os.path.dirname(__file__))
+
 @pt.fixture
 def app() -> "fe.FlaskExpress":
     app = fe.FlaskExpress(__name__)
     app.secret_key = "top-secret-key"
     app.testing = True
-    app.template_folder = os.path.join(os.path.abspath(os.path.dirname(__file__)), 'files')
-    app.config['ATTACHMENTS_FOLDER'] = os.path.join(os.path.abspath(os.path.dirname(__file__)), 'files')
+    app.template_folder = os.path.join(BASE_DIR, 'files')
+    app.config['ATTACHMENTS_FOLDER'] = os.path.join(BASE_DIR, 'files')
     return app