+++ CHANGELOG.md
@@ -17,3 +17,4 @@
 - Now the user have the power to add custom attachment folder.
 - showing proper error message when attached file is absent at the attachment directory.
 - added more CamelCase method to the `response.Response` class.
+- fixed `issubclass` checking error for different contents at `response.Response.send` method.